Move l.schema to l.lib.schema.
